+++ main.py
@@ -23,7 +23,9 @@ async def load_cogs():
         await bot.load_extension('commands.credits')
         await bot.load_extension('commands.users')
         await bot.load_extension('commands.clearvideos')
-        await bot.load_extension('commands.helpc')
+        await bot.load_extension('commands.helpc')  
+        await bot.load_extension('commands.tables')
+        await bot.load_extension('commands.viewer')
 
     except Exception as e:
         print(f'Failed to load cogs: {e}')